SQL_AGENT/requirements.txt
langchain
langgraph  
langchain-community
langchain-classic
langchain-google-genai
langchain-openai
sqlalchemy
requests
langsmith
langgraph-checkpoint
pydantic
pydantic-settings
dotenv
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
motor 
pymongo
